Refactor out common SVG + HTML element definition code
