+++ lib/Document.js
@@ -161,12 +161,12 @@ Document.prototype = Object.create(Node.prototype, {
        qualifiedName !== 'xmlns' &&
        prefix !== 'xmlns'))
       utils.NamespaceError();
-      
+
     if (namespace === NAMESPACE.HTML) {
       return html.createElement(this, localName, prefix);
     }
     else if (namespace === NAMESPACE.SVG) {
-      return svg.createElement(this, localName, namespace, prefix);
+      return svg.createElement(this, localName, prefix);
     }
 
     return new Element(this, localName, namespace, prefix);
