+++ lib/htmlelts.js
@@ -3,77 +3,19 @@ var Node = require('./Node');
 var Element = require('./Element');
 var CSSStyleDeclaration = require('./CSSStyleDeclaration');
 var utils = require('./utils');
-var attributes = require('./attributes');
 var URLUtils = require('./URLUtils');
-var sloppy = require('./sloppy');
+var defineElement = require('./defineElement');
 
-var impl = exports.elements = {};
-var tagNameToImpl = Object.create(null);
+var htmlElements = exports.elements = {};
+var htmlNameToImpl = Object.create(null);
 
 exports.createElement = function(doc, localName, prefix) {
-  var impl = tagNameToImpl[localName] || HTMLUnknownElement;
+  var impl = htmlNameToImpl[localName] || HTMLUnknownElement;
   return new impl(doc, localName, prefix);
 };
 
 function define(spec) {
-  var c = spec.ctor;
-  if (c) {
-    var props = spec.props || {};
-    if (spec.attributes) {
-      for (var n in spec.attributes) {
-        var attr = spec.attributes[n];
-        if (typeof attr !== 'object' || Array.isArray(attr)) attr = {type: attr};
-        if (!attr.name) attr.name = n.toLowerCase();
-        props[n] = attributes.property(attr);
-      }
-    }
-    props.constructor = { value : c };
-    c.prototype = Object.create((spec.superclass || HTMLElement).prototype, props);
-    if (spec.events) {
-      addEventHandlers(c, spec.events);
-    }
-    impl[c.name] = c;
-  }
-  else {
-    c = HTMLElement;
-  }
-  (spec.tags || spec.tag && [spec.tag] || []).forEach(function(tag) {
-    tagNameToImpl[tag] = c;
-  });
-  return c;
-}
-
-function EventHandlerBuilder(body, document, form, element) {
-  this.body = body;
-  this.document = document;
-  this.form = form;
-  this.element = element;
-}
-
-EventHandlerBuilder.prototype.build = sloppy.EventHandlerBuilder_build;
-
-function EventHandlerChangeHandler(elt, name, oldval, newval) {
-  var doc = elt.ownerDocument || Object.create(null);
-  var form = elt.form || Object.create(null);
-  elt[name] = new EventHandlerBuilder(newval, doc, form, elt).build();
-}
-
-function addEventHandlers(c, eventHandlerTypes) {
-  var p = c.prototype;
-  eventHandlerTypes.forEach(function(type) {
-    // Define the event handler registration IDL attribute for this type
-    Object.defineProperty(p, "on" + type, {
-      get: function() {
-        return this._getEventHandler(type);
-      },
-      set: function(v) {
-        this._setEventHandler(type, v);
-      },
-    });
-
-    // Define special behavior for the content attribute as well
-    attributes.registerChangeHandler(c, "on" + type, EventHandlerChangeHandler);
-  });
+  return defineElement(spec, HTMLElement, htmlElements, htmlNameToImpl);
 }
 
 function URL(attr) {
@@ -114,7 +56,7 @@ var HTMLElement = exports.HTMLElement = define({
         parser.parse(v, true);
         var tmpdoc = parser.document();
         var root = tmpdoc.firstChild;
-        var target = (this instanceof tagNameToImpl.template) ?
+        var target = (this instanceof htmlNameToImpl.template) ?
             this.content : this;
 
         // Remove any existing children of this node
@@ -266,7 +208,7 @@ define({
     type: String
   }
 });
-URLUtils._inherit(tagNameToImpl.a.prototype);
+URLUtils._inherit(htmlNameToImpl.a.prototype);
 
 define({
   tag: 'area',
@@ -1055,17 +997,17 @@ define({
 
 define({
   tag: 'audio',
-  superclass: impl.HTMLMediaElement,
+  superclass: htmlElements.HTMLMediaElement,
   ctor: function HTMLAudioElement(doc, localName, prefix) {
-    impl.HTMLMediaElement.call(this, doc, localName, prefix);
+    htmlElements.HTMLMediaElement.call(this, doc, localName, prefix);
   }
 });
 
 define({
   tag: 'video',
-  superclass: impl.HTMLMediaElement,
+  superclass: htmlElements.HTMLMediaElement,
   ctor: function HTMLVideoElement(doc, localName, prefix) {
-    impl.HTMLMediaElement.call(this, doc, localName, prefix);
+    htmlElements.HTMLMediaElement.call(this, doc, localName, prefix);
   },
   attributes: {
     poster: String,
@@ -1076,17 +1018,17 @@ define({
 
 define({
   tag: 'td',
-  superclass: impl.HTMLTableCellElement,
+  superclass: htmlElements.HTMLTableCellElement,
   ctor: function HTMLTableDataCellElement(doc, localName, prefix) {
-    impl.HTMLTableCellElement.call(this, doc, localName, prefix);
+    htmlElements.HTMLTableCellElement.call(this, doc, localName, prefix);
   }
 });
 
 define({
   tag: 'th',
-  superclass: impl.HTMLTableCellElement,
+  superclass: htmlElements.HTMLTableCellElement,
   ctor: function HTMLTableHeaderCellElement(doc, localName, prefix) {
-    impl.HTMLTableCellElement.call(this, doc, localName, prefix);
+    htmlElements.HTMLTableCellElement.call(this, doc, localName, prefix);
   },
   attributes: {
     scope: ["", "row", "col", "rowgroup", "colgroup"]
