+++ lib/svg.js
@@ -1,44 +1,24 @@
 "use strict";
 var Element = require('./Element');
-var attributes = require('./attributes');
+var defineElement = require('./defineElement');
+var utils = require('./utils');
 var CSSStyleDeclaration = require('./CSSStyleDeclaration');
 
-var impl = exports.elements = {};
-var tagNameToImpl = Object.create(null);
+var svgElements = exports.elements = {};
+var svgNameToImpl = Object.create(null);
 
-exports.createElement = function(doc, localName, namespaceURI, prefix) {
-  var elementImpl = tagNameToImpl[localName] || SVGElement;
-  return new elementImpl(doc, localName, namespaceURI, prefix);
+exports.createElement = function(doc, localName, prefix) {
+  var impl = svgNameToImpl[localName] || SVGElement;
+  return new impl(doc, localName, prefix, utils.NAMESPACE.SVG);
 };
 
 function define(spec) {
-  var c = spec.ctor;
-  if (c) {
-    var props = spec.props || {};
-    if (spec.attributes) {
-      for (var n in spec.attributes) {
-        var attr = spec.attributes[n];
-        if (typeof attr !== 'object' || Array.isArray(attr)) attr = {type: attr};
-        if (!attr.name) attr.name = n.toLowerCase();
-        props[n] = attributes.property(attr);
-      }
-    }
-    props.constructor = { value : c };
-    c.prototype = Object.create((spec.superclass || SVGElement).prototype, props);
-    impl[c.name] = c;
-  }
-  else {
-    c = SVGElement;
-  }
-  (spec.tags || spec.tag && [spec.tag] || []).forEach(function(tag) {
-    tagNameToImpl[tag] = c;
-  });
-  return c;
+  return defineElement(spec, SVGElement, svgElements, svgNameToImpl);
 }
 
-var SVGElement = exports.SVGElement = define({
+var SVGElement = define({
   superclass: Element,
-  ctor: function SVGElement(doc, localName, namespaceURI, prefix) {
+  ctor: function SVGElement(doc, localName, prefix, namespaceURI) {
     Element.call(this, doc, localName, namespaceURI, prefix);
   },
   props: {
@@ -51,13 +31,13 @@ var SVGElement = exports.SVGElement = define({
 });
 
 define({
-  ctor: function SVGSVGElement(doc, localName, namespaceURI, prefix) {
-    SVGElement.call(this, doc, localName, namespaceURI, prefix);
+  ctor: function SVGSVGElement(doc, localName, prefix, namespaceURI) {
+    SVGElement.call(this, doc, localName, prefix, namespaceURI);
   },
   tag: 'svg',
   props: {
     createSVGRect: { value: function () {
-      return new SVGElement(this.ownerDocument, 'rect', this.namespaceURI, this.prefix);
+      return new SVGElement(this.ownerDocument, 'rect', this.prefix, this.namespaceURI);
     } }
   }
 });
